+++ test/LionWeb-CSharp-Test/tests/SerializationTests.cs
@@ -51,6 +51,100 @@ public class SerializationTests
         new Deserializer([_language]).Deserialize(serializationChunk);
     }
 
+    // Disabled until #19 is merged
+    // [TestMethod]
+    // public void test_serialization_shapes_language()
+    // {
+    //     var serializationChunk = LanguageSerializer.Serialize(ShapesLanguage.Instance);
+    //     Console.WriteLine(JsonUtils.WriteJsonToString(serializationChunk));
+    //
+    //     var redeserialized = LanguageDeserializer.Deserialize(serializationChunk);
+    //     var comparer = new Comparer([ShapesLanguage.Instance], redeserialized.Cast<IReadableNode>().ToList());
+    //     Assert.IsTrue(comparer.AreEqual(), comparer.ToMessage(new ComparerOutputConfig()));
+    // }
+
+    [TestMethod]
+    public void test_serialization_shapes_language_with_external_annotations()
+    {
+        var serializationChunk = LanguageSerializer.Serialize(ShapesLanguage.Instance);
+
+        var redeserialized = LanguageDeserializer.Deserialize(serializationChunk);
+        Language redeserializedShapes = redeserialized.Cast<INode>().OfType<Language>().First();
+
+        var language = new DynamicLanguage("id-anns") { Key = "key-anns", Name = "Anns", Version = "1" };
+
+        var ann = language.Annotation("id-classifierAnn", "key-classifierAnn", "ClassifierAnn");
+        Property age = ann
+            .Annotating(M3Language.Instance.Classifier)
+            .Implementing(BuiltInsLanguage.Instance.INamed)
+            .Property("id-age", "key-age", "Age").OfType(BuiltInsLanguage.Instance.Integer);
+
+        DynamicClassifier redeserializedCircle =
+            (DynamicClassifier)redeserializedShapes.ClassifierByKey(ShapesLanguage.Instance.Circle.Key);
+        var annInst = language.GetFactory().CreateNode("ann-inst", ann);
+        annInst.Set(BuiltInsLanguage.Instance.INamed_name, "Hello");
+        annInst.Set(age, 23);
+        redeserializedCircle.AddAnnotations([annInst]);
+
+        var serializationChunk2 = LanguageSerializer.Serialize(redeserializedShapes);
+        Console.WriteLine(JsonUtils.WriteJsonToString(serializationChunk2));
+        var redeserialized2 = LanguageDeserializer.Deserialize(serializationChunk2, language);
+        Language redeserializedShapes2 = redeserialized2.Cast<INode>().OfType<Language>().First();
+        DynamicClassifier redeserializedCircle2 =
+            (DynamicClassifier)redeserializedShapes2.ClassifierByKey(ShapesLanguage.Instance.Circle.Key);
+
+        var comparer = new Comparer(redeserializedCircle.GetAnnotations().ToList(),
+            redeserializedCircle2.GetAnnotations().ToList());
+        Assert.IsTrue(comparer.AreEqual(), comparer.ToMessage(new ComparerOutputConfig()));
+    }
+
+    [TestMethod]
+    public void test_serialization_shapes_language_together_with_annotations()
+    {
+        var serializationChunk = LanguageSerializer.Serialize(ShapesLanguage.Instance);
+
+        var redeserialized = LanguageDeserializer.Deserialize(serializationChunk);
+        Language redeserializedShapes = redeserialized.Cast<INode>().OfType<Language>().First();
+
+        var language = new DynamicLanguage("id-anns") { Key = "key-anns", Name = "Anns", Version = "1" };
+
+        var ann = language.Annotation("id-classifierAnn", "key-classifierAnn", "ClassifierAnn");
+        Property age = ann
+            .Annotating(M3Language.Instance.Classifier)
+            .Implementing(BuiltInsLanguage.Instance.INamed)
+            .Property("id-age", "key-age", "Age").OfType(BuiltInsLanguage.Instance.Integer);
+
+        DynamicClassifier redeserializedCircle =
+            (DynamicClassifier)redeserializedShapes.ClassifierByKey(ShapesLanguage.Instance.Circle.Key);
+        var annInst = language.GetFactory().CreateNode("ann-inst", ann);
+        annInst.Set(BuiltInsLanguage.Instance.INamed_name, "Hello");
+        annInst.Set(age, 23);
+        redeserializedCircle.AddAnnotations([annInst]);
+
+        var serializationChunk2 = LanguageSerializer.Serialize(redeserializedShapes, language);
+        Console.WriteLine(JsonUtils.WriteJsonToString(serializationChunk2));
+        var redeserialized2 = LanguageDeserializer.Deserialize(serializationChunk2);
+        Language redeserializedShapes2 = redeserialized2.Cast<INode>().OfType<Language>()
+            .First(l => l.Key == ShapesLanguage.Instance.Key);
+        DynamicClassifier redeserializedCircle2 =
+            (DynamicClassifier)redeserializedShapes2.ClassifierByKey(ShapesLanguage.Instance.Circle.Key);
+
+        var comparer = new Comparer(redeserializedCircle.GetAnnotations().ToList(),
+            redeserializedCircle2.GetAnnotations().ToList());
+        Assert.IsTrue(comparer.AreEqual(), comparer.ToMessage(new ComparerOutputConfig()));
+    }
+
+    [TestMethod]
+    public void test_serialization_lioncore()
+    {
+        var serializationChunk = LanguageSerializer.Serialize(M3Language.Instance);
+        Console.WriteLine(JsonUtils.WriteJsonToString(serializationChunk));
+
+        // Just run the deserializer for now (without really checking anything), to see whether it crashes or not:
+        var deserializer = new LanguageDeserializer(serializationChunk, false);
+        deserializer.DeserializeLanguages();
+    }
+
     [TestMethod]
     public void test_serialization_shapes_language()
     {
@@ -76,7 +170,8 @@ public class SerializationTests
     public void test_no_double_serialization()
     {
         var geometry = ExampleModels.ExampleModel(_language);
-        var shape0 = (geometry.Get(_language.ClassifierByKey("key-Geometry").FeatureByKey("key-shapes")) as IEnumerable).Cast<INode>().First();
+        var shape0 = (geometry.Get(_language.ClassifierByKey("key-Geometry").FeatureByKey("key-shapes")) as IEnumerable)
+            .Cast<INode>().First();
 
         Assert.IsInstanceOfType<INode>(shape0);
         var serializationChunk = Serializer.Serialize([geometry, shape0]);
@@ -111,7 +206,7 @@ public class SerializationTests
     public void SerializeUnsetRequiredContainment()
     {
         var compositeShape = new CompositeShape("comp");
-        
+
         var serializationChunk = Serializer.Serialize([compositeShape]);
         var nodes = new Deserializer([ShapesLanguage.Instance]).Deserialize(serializationChunk);